rename mapper and dependencies update
